+++ lib/controllers/controllers.dart
@@ -2,5 +2,4 @@
 // Controller-related classes and utilities
 
 export 'zen_controller.dart' show ZenController;
-export 'zen_route_observer.dart';
 export 'zen_service.dart';