+++ lib/controllers/zen_controller.dart
@@ -874,4 +874,3 @@ extension ZenControllerWorkerExtension on ZenController {
     }
   }
 }
-