+++ test/di/zen_service_integration_test.dart
@@ -20,7 +20,6 @@ class TestController extends ZenController {
 void main() {
   group('ZenService Integration', () {
     tearDown(() {
-      ZenService.disposeAllServices();
       Zen.reset();
     });